param renamed to param_change
This way the filename matches the class name.
